requirements.txt
aiogram
aioschedule
python-dotenv
